Exercices before Django

+++ Exercicio3 - calculo salario liquido.py
@@ -1,5 +1,5 @@
 valor_hora_aula = float(input("Informe o seu valor por hora: "))
-numero_horas = float(input("Informe o número de horas aula: "))
+numero_horas = int(input("Informe o número de horas aula: "))
 descontos = float(input("Informe os descontos: "))
 
 salario_liquido = (valor_hora_aula * numero_horas) - descontos
